chore: remove unnecessary clones and conversions
Address clippy warnings in dns, routing, and tun modules.

+++ src/tun.rs
@@ -294,7 +294,7 @@ pub(crate) fn spawn_writer<T: TunTx>(
                     }
 
                     match tun.send(&packet).await {
-                        Ok(written) => counters.record_success(written as usize),
+                        Ok(written) => counters.record_success(written),
                         Err(err) if err.kind() == io::ErrorKind::Interrupted => continue,
                         Err(_) => break,
                     }
